requirements.txt
fastapi
uvicorn
chromadb
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-groq
huggingface-hub
sentence-transformers
torch
pymupdf
pydantic
python-dotenv
